Add option to sync NetCDF files
Sometimes, NetCDF files need syncing (force-writing to disk). This seems
to be particularly the case for GPU runs.

This PR adds an option to provide the `NetCDFWriter` with a schedule to
call `NCDatasets.sync` based on arbitrary conditions.

The default behavior depends on the context: on CPUs, we let `NetCDF`
manage its buffered writes, on GPUs, we call `sync` at the every time
steps for those datasets that need to be synced (ie, those that were
written since the last sync).

In general, we keep track of what files have been recently
touched (adding them to a list internal to the `writer`) and whenever
`sync` is called, `NCDatasets.sync` is called on those files. This
ensures that only files that needs to be synced are synced and that
calling `sync` twice in a row results in a no-op for the second `sync`.