Amend netcdf_writer.jl
This is a small PR removing 3 lines of code from netcdf_writer.jl,
these 3 lines are already inside a loop checking if space is horizontal,
and need to be evaluated only if it is not (if space is 3D).
In the case where space is 3D, we have a z coordinate, if space is
horizontal, we do not.