Fix interpolation of two diagnostics with same short_name
NetCDFWriter was internally storing preallocated spaces in a dictionary. The
keys of this dictionary were the short names of output variable, but these are
not unique across diagnostics. For instance, two diagnostics computed with
period 2 and one with period 3 are distinct but share the same short name. When
multiple distinct diagnostics shared the same name, the values written to file
would not be correct. This commit fixes this bug by directly using the
diagnostic as key.