Introduce z_sampling_method, remove disable_vertical_interpolation
The NetCDF writer was originally designed for ClimaAtmos. As such, it uses some
fake pressure levels for vertical sampling as default. This is generally not
applicable/useful. This commit generalized the method two sample points on the
vertical direction introducing a new type `AbstractZSamplingMethod`.

`AbstractZSamplingMethod` has to concrete subtypes:
- `LevelsMethod` (equivalent to `disable_vertical_interpolation`)
- `FakePressureLevelsMethod` (equivalent to previously default behavior)

This opens up the possibility to also define `LinearMethod`, for linear
sampling, and `LogMethod`, for log sampling.

The default was changed to `LevelsMethod` and `disable_vertical_interpolation`
was deprecated.