Add NetCDFWriter support for long-lat boxes
This commit makes identifying coordinates for resampling more
coordinate agnostic.

This commit introduces a workaround for
https://github.com/CliMA/ClimaCore.jl/issues/1936

More fundamentally, https://github.com/CliMA/ClimaCore.jl/issues/1936 is
probably a reflection of the fact that boxes with longlat points are not
supported extremely well.

Here, we make the decision to work with LatLong boxes (instead of the
more natural LongLat ones, more natural because Long is the X axis).
LongLat boxes will fail