Save ScheduleDiagnostics as Vectors
`ScheduleDiagnostic`s with `NetCDFWriter`s have a complex type structure
that can lead to significant compile time. We originally saved
`ScheduleDiagnostic`s internally as tuples to improve inferrability, but
this does not seem to scale well with lots of diagnostics.

In the future, we might want to consider looking at the type signature
and move the reference to the writer outside of `ScheduleDiagnostic`s
and into the `DiagnosticHandler`.