Fix bug with box domain
When the user pass in a value for `num_points` when constructing the
`NetCDFWriter`, the first and second elements of a tuple represent
the number of points for the latitude and longitude dimensions instead
of the number of points for the longitude and latitude dimensions. This
commit changes the behavior of `target_coordinates` to always to return
the order as (longitude, latitude, vertical) or (x, y, vertical).